[AMRS-199] - Updated - returning correct status to other sections

+++ ai_mf_backend/models/v1/api/questionnaire.py
@@ -55,7 +55,7 @@ class SectionCompletionStatus(BaseModel):
     section_name: str
     answered_questions: int
     total_questions: int
-    completion_rate: float
+    completion_rate: int
 
 
 class SectionCompletionStatusResponse(Response):
